Add placement validations concern
